Move ADO-related functions out of commonHelper
