+++ src/configure/helper/commonHelper.ts
@@ -10,38 +10,6 @@ export async function sleepForMilliSeconds(timeInMs: number): Promise<void> {
     });
 }
 
-export function generateDevOpsOrganizationName(userName: string, repositoryName: string): string {
-    let repositoryNameSuffix = repositoryName.replace("/", "-").trim();
-    let organizationName = `${userName}-${repositoryNameSuffix}`;
-
-    // Name cannot start or end with whitespaces, cannot start with '-', cannot contain characters other than a-z|A-Z|0-9
-    organizationName = organizationName.trim().replace(/^[-]+/, '').replace(/[^a-zA-Z0-9-]/g, '');
-    if(organizationName.length > 50) {
-        organizationName = organizationName.substr(0, 50);
-    }
-
-    return organizationName;
-}
-
-export function generateDevOpsProjectName(repositoryName?: string): string {
-    if(!repositoryName) {
-        return "AzurePipelines";
-    }
-
-    let repoParts = repositoryName.split("/");
-    let suffix = repoParts[repoParts.length-1];
-    suffix = suffix.trim();
-    // project name cannot end with . or _
-    suffix = suffix.replace(/\.[\.]*$/, '').replace(/^_[_]*$/, '');
-
-    let projectName = `AzurePipelines-${suffix}`;
-    if(projectName.length > 64) {
-        projectName = projectName.substr(0, 64);
-    }
-
-    return projectName;
-}
-
 export function generateRandomPassword(length: number = 20): string {
     var characters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890!@#%^*()-+";
     var charTypeSize = new Array(26, 26, 10, 10);
